requirements.txt
gunicorn
pymongo
python-dotenv
azure-storage-blob
motor
fastapi
uvicorn
python-multipart
passlib
bcrypt
dnspython
httpx
emails
pydantic[email]
PyJWT
python-jose[cryptography]
email-validator
